GH-43168: [Swift] Add buffer and array builders for Struct type (#43171)
### Rationale for this change
The struct type requires additions/modifications to builders for buffers and arrays.   This is needed in order match functionality with the other supported types.

### What changes are included in this PR?
Updates to buffer and array builders for Struct type.

### Are these changes tested?
Yes, Array test has been added for the Struct type using the Struct builders

**This PR includes breaking changes to public APIs.**
The init method for ArrowArray currently does not throw.  This PR will change this behavior and allow the init method to throw.  External code that calls init will need to be updated to catch any errors. 

* GitHub Issue: #43168

Authored-by: Alva Bandy <abandy@live.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>