GH-46296: [Swift] Add support for reading struct (#46302)
### Rationale for this change

`makeArrayHolder` supports `ArrowTypeId.strct`.

https://github.com/apache/arrow/blob/62c59c23352583980e894a43ec46335c4d55a7e1/swift/Arrow/Sources/Arrow/ArrowReaderHelper.swift#L189-L190

However, we cannot use `makeArrayHolder` if `fromProto` fails to handle `org_apache_arrow_flatbuf_Type_.struct_`.

https://github.com/apache/arrow/blob/62c59c23352583980e894a43ec46335c4d55a7e1/swift/Arrow/Sources/Arrow/ArrowReaderHelper.swift#L139-L148

### What changes are included in this PR?

This PR aims to fix `ProtoUtil.fromProto` to support `struct_`.

This PR also adds an unit test for it.

This is based on GH-46297.

### Known issues with Swift Arrow Reader

Currently the Swift Arrow reader requires a null buffer to be included in the list of buffers.  The reader will fail if the null buffer is missing.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.

* GitHub Issue: #46296

Co-authored-by: Dongjoon Hyun <dongjoon@ apache.org>

Authored-by: Alva Bandy <abandy@live.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>