fix: Padding now added to Arrow file marker and RecordBatches are being written with correct alignment (#95)
Files were being written with a non-padded file marker and alignment was
not being written when serializing record batches. Also the metadata
length was being set to zero in the block.

## What's Changed

The padded version of the filemarker is being written.

Padding is written to record blocks and record block metadata, i.e.:
```
 addPadForAlignment(&writer)
```

The metadata length was being written as zero. This was preventing
PyArrow from reading files written by `ArrowWriter`. This has now been
calculated and set in the Block.

Closes #91.